Resolving cds.requires.llm early on and by default in dev mode
